core: move tests from src to test
