+++ packages/core/test/lib/validators.test.ts
@@ -1,5 +1,5 @@
-import intRange from '../validators/int-range'
-import stringWithLength from '../validators/string-with-length'
+import intRange from '../../src/lib/validators/int-range'
+import stringWithLength from '../../src/lib/validators/string-with-length'
 
 describe('intRange(min, max)(val)', () => {
   it('work with various values', () => {
